client/requirements.txt
pycrypto==2.6.1
requests==2.2.1
Pillow==2.1.0
stepic3==0.4